core-app-api: moved out and delay the route binding logic
