+++ packages/core-app-api/src/app/AppManager.tsx
@@ -45,9 +45,6 @@ import {
   IdentityApi,
   identityApiRef,
   BackstagePlugin,
-  RouteRef,
-  SubRouteRef,
-  ExternalRouteRef,
 } from '@backstage/core-plugin-api';
 import { ApiFactoryRegistry, ApiResolver } from '../apis/system';
 import {
@@ -75,48 +72,19 @@ import {
   AppConfigLoader,
   AppContext,
   AppOptions,
-  AppRouteBinder,
   BackstageApp,
   SignInPageProps,
 } from './types';
 import { AppThemeProvider } from './AppThemeProvider';
 import { defaultConfigLoader } from './defaultConfigLoader';
 import { ApiRegistry } from '../apis/system/ApiRegistry';
+import { resolveRouteBindings } from './resolveRouteBindings';
 
 type CompatiblePlugin =
   | BackstagePlugin<any, any>
   | (Omit<BackstagePlugin<any, any>, 'getFeatureFlags'> & {
       output(): Array<{ type: 'feature-flag'; name: string }>;
     });
-
-export function generateBoundRoutes(bindRoutes: AppOptions['bindRoutes']) {
-  const result = new Map<ExternalRouteRef, RouteRef | SubRouteRef>();
-
-  if (bindRoutes) {
-    const bind: AppRouteBinder = (
-      externalRoutes,
-      targetRoutes: { [name: string]: RouteRef | SubRouteRef },
-    ) => {
-      for (const [key, value] of Object.entries(targetRoutes)) {
-        const externalRoute = externalRoutes[key];
-        if (!externalRoute) {
-          throw new Error(`Key ${key} is not an existing external route`);
-        }
-        if (!value && !externalRoute.optional) {
-          throw new Error(
-            `External route ${key} is required but was undefined`,
-          );
-        }
-        if (value) {
-          result.set(externalRoute, value);
-        }
-      }
-    };
-    bindRoutes({ bind });
-  }
-
-  return result;
-}
 
 /**
  * Get the app base path from the configured app baseUrl.
@@ -196,7 +164,7 @@ export class AppManager implements BackstageApp {
   private readonly themes: AppTheme[];
   private readonly configLoader?: AppConfigLoader;
   private readonly defaultApis: Iterable<AnyApiFactory>;
-  private readonly routeBindings: Map<ExternalRouteRef, RouteRef | SubRouteRef>;
+  private readonly bindRoutes: AppOptions['bindRoutes'];
 
   private readonly appIdentityProxy = new AppIdentityProxy();
   private readonly apiFactoryRegistry: ApiFactoryRegistry;
@@ -209,7 +177,7 @@ export class AppManager implements BackstageApp {
     this.themes = options.themes as AppTheme[];
     this.configLoader = options.configLoader ?? defaultConfigLoader;
     this.defaultApis = options.defaultApis ?? [];
-    this.routeBindings = generateBoundRoutes(options.bindRoutes);
+    this.bindRoutes = options.bindRoutes;
     this.apiFactoryRegistry = new ApiFactoryRegistry();
   }
 
@@ -237,37 +205,45 @@ export class AppManager implements BackstageApp {
         [],
       );
 
-      const { routePaths, routeParents, routeObjects, featureFlags } =
-        useMemo(() => {
-          const result = traverseElementTree({
-            root: children,
-            discoverers: [childDiscoverer, routeElementDiscoverer],
-            collectors: {
-              routePaths: routePathCollector,
-              routeParents: routeParentCollector,
-              routeObjects: routeObjectCollector,
-              collectedPlugins: pluginCollector,
-              featureFlags: featureFlagCollector,
-            },
-          });
+      const {
+        routePaths,
+        routeParents,
+        routeObjects,
+        featureFlags,
+        routeBindings,
+      } = useMemo(() => {
+        const result = traverseElementTree({
+          root: children,
+          discoverers: [childDiscoverer, routeElementDiscoverer],
+          collectors: {
+            routePaths: routePathCollector,
+            routeParents: routeParentCollector,
+            routeObjects: routeObjectCollector,
+            collectedPlugins: pluginCollector,
+            featureFlags: featureFlagCollector,
+          },
+        });
 
-          // TODO(Rugvip): Restructure the public API so that we can get an immediate view of
-          //               the app, rather than having to wait for the provider to render.
-          //               For now we need to push the additional plugins we find during
-          //               collection and then make sure we initialize things afterwards.
-          result.collectedPlugins.forEach(plugin => this.plugins.add(plugin));
-          this.verifyPlugins(this.plugins);
+        // TODO(Rugvip): Restructure the public API so that we can get an immediate view of
+        //               the app, rather than having to wait for the provider to render.
+        //               For now we need to push the additional plugins we find during
+        //               collection and then make sure we initialize things afterwards.
+        result.collectedPlugins.forEach(plugin => this.plugins.add(plugin));
+        this.verifyPlugins(this.plugins);
 
-          // Initialize APIs once all plugins are available
-          this.getApiHolder();
-          return result;
-        }, [children]);
+        // Initialize APIs once all plugins are available
+        this.getApiHolder();
+        return {
+          ...result,
+          routeBindings: resolveRouteBindings(this.bindRoutes),
+        };
+      }, [children]);
 
       if (!routesHaveBeenValidated) {
         routesHaveBeenValidated = true;
         validateRouteParameters(routePaths, routeParents);
         validateRouteBindings(
-          this.routeBindings,
+          routeBindings,
           this.plugins as Iterable<BackstagePlugin<any, any>>,
         );
       }
@@ -331,7 +307,7 @@ export class AppManager implements BackstageApp {
                 routePaths={routePaths}
                 routeParents={routeParents}
                 routeObjects={routeObjects}
-                routeBindings={this.routeBindings}
+                routeBindings={routeBindings}
                 basePath={getBasePath(loadedConfig.api)}
               >
                 {children}
